// src/types/fontTypes.ts
export enum FontWeight {
  Thin = '100',
  ExtraLight = '200',
  Light = '300',
  Normal = '400',
  Medium = '500',
  SemiBold = '600',
  Bold = '700',
  ExtraBold = '800',
  Black = '900',
}

export enum FontSize {
  XsSmallest = 6,
  XsSmall = 8,
  ExtraSmall = 10,
  Small = 12,
  MediumSmall = 14,
  Medium = 16,
  MediumLarge = 18,
  Large = 20,
  ExtraLarge = 24,
  ExtraExtraLarge = 28,
  XL = 30,
  XXL = 36,
  XXXL = 42,
  Huge = 48,
  Enormous = 60,
}